start changing info to sidebar
