+++ src/components/PhotoViewer/FileInfo/ExifData.tsx
