+++ src/components/PhotoViewer/FileInfo/RenderCreationTime.tsx
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 import { updateFilePublicMagicMetadata } from 'services/fileService';
 import { EnteFile } from 'types/file';
-import constants from 'utils/strings/constants';
+import CalendarTodayIcon from '@mui/icons-material/CalendarToday';
 import {
     changeFileCreationTime,
     updateExistingFilePubMetadata,
 } from 'utils/file';
 import { formatDateTime } from 'utils/time';
 import EditIcon from '@mui/icons-material/Edit';
-import { Label, Row, Value } from 'components/Container';
+import { FlexWrapper, Value } from 'components/Container';
 import { logError } from 'utils/sentry';
 import { SmallLoadingSpinner } from '../styledComponents/SmallLoadingSpinner';
 import EnteDateTimePicker from 'components/EnteDateTimePicker';
@@ -58,40 +58,35 @@ export function RenderCreationTime({
     };
 
     return (
-        <>
-            <Row>
-                <Label width="30%">{constants.CREATION_TIME}</Label>
+        <FlexWrapper>
+            <CalendarTodayIcon />
+            <Value width={!shouldDisableEdits ? !isInEditMode && '60%' : '70%'}>
+                {isInEditMode ? (
+                    <EnteDateTimePicker
+                        initialValue={originalCreationTime}
+                        disabled={loading}
+                        onSubmit={saveEdits}
+                        onClose={closeEditMode}
+                    />
+                ) : (
+                    formatDateTime(originalCreationTime)
+                )}
+            </Value>
+            {!shouldDisableEdits && !isInEditMode && (
                 <Value
-                    width={
-                        !shouldDisableEdits ? !isInEditMode && '60%' : '70%'
-                    }>
-                    {isInEditMode ? (
-                        <EnteDateTimePicker
-                            initialValue={originalCreationTime}
-                            disabled={loading}
-                            onSubmit={saveEdits}
-                            onClose={closeEditMode}
-                        />
+                    width={'10%'}
+                    style={{ cursor: 'pointer', marginLeft: '10px' }}>
+                    {loading ? (
+                        <IconButton>
+                            <SmallLoadingSpinner />
+                        </IconButton>
                     ) : (
-                        formatDateTime(originalCreationTime)
+                        <IconButton onClick={openEditMode}>
+                            <EditIcon />
+                        </IconButton>
                     )}
                 </Value>
-                {!shouldDisableEdits && !isInEditMode && (
-                    <Value
-                        width={'10%'}
-                        style={{ cursor: 'pointer', marginLeft: '10px' }}>
-                        {loading ? (
-                            <IconButton>
-                                <SmallLoadingSpinner />
-                            </IconButton>
-                        ) : (
-                            <IconButton onClick={openEditMode}>
-                                <EditIcon />
-                            </IconButton>
-                        )}
-                    </Value>
-                )}
-            </Row>
-        </>
+            )}
+        </FlexWrapper>
     );
 }
