+++ src/components/PhotoViewer/FileInfo/RenderInfoItem.tsx
