+++ src/utils/strings/englishConstants.tsx
@@ -436,11 +436,13 @@ const englishConstants = {
     FILE_ID: 'File ID',
     FILE_NAME: 'File name',
     CAPTION: 'Description',
-    CAPTION_PLACEHOLDER: 'Add a description',
+    CAPTION_PLACEHOLDER: 'Add a description... ',
     CREATION_TIME: 'Creation time',
     UPDATED_ON: 'Updated on',
     LOCATION: 'Location',
-    SHOW_MAP: 'Show on map',
+    SHOW_ON_MAP: 'View on OpenStreetMap',
+    DETAILS: 'Details',
+    VIEW_EXIF: 'View all EXIF data',
     EXIF: 'Exif',
     DEVICE: 'Device',
     IMAGE_SIZE: 'Image size',
